pyproject.toml
[tool.poetry]
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
gunicorn = "^23.0.0"
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
celery = {extras = ["librabbitmq", "redis"], version = "^5.4.0"}
redis = "^5.2.0"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.2"
pytest = "^8.3.3"
pylint = "^3.3.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
celery-stubs = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# black
[tool.black]
line-length = 79
extend-exclude = "migrations/"

# isort
[tool.isort]
profile = "black"
extend_skip_glob = ["migrations"]
known_local_folder=["api"]
line_length = 79
lines_after_imports = 2

# mypy
[tool.mypy]
exclude = ["migrations/"]

[tool.pylint]
ignore = [".git", "__pycache__", "migrations", ".venv"]
max-line-length = 79
disable = ["C0114", "C0115", "C0116", "W1514", "W3101", "R0903", "C0103"]
